Factor out resource retrieval/task storage (#393)
